Latest work from Jason, need to still update CHANGELOG, but this is what he wrote:
Here is my latest set of diffs. I figured I have
enough here to send it to you.  This diff has the
following changes.

1. A New option to turn off comment email. We have
already discussed this.
2. codestriker.pl, when run with no url params, shows
a list of topics that are open. Before it would show
all topics.
3. fix: full email addresses where being displayed in
the show all comment page when spam email support was
configured.
4. fix: spelling mistake in javascript help in
createtopic page.
5. Performance improvement. Changed the
codestriker.pl, to not load all of the .pm files at
the start of the file. It will only load pm files when
needed. This avoids perl from loading and compiling
the entire code base on every request. This small
changed, helps everything go faster.

Any the big change....

6. The view all comments window has three combo boxs
at the top of the file that allow the user to filter
by author name, by state, and to optionally show
context or not show context in the comment. This
should make it  faster for reviewers to prepare for
meetings, and faster for topic authors to manage the
comments after the meeting. From the user view
perspective, I am really happy with how this came out.