Modified viewtopicheader so that codestriker.css is used, and so that the tabbing works for Mozilla, Netscape and Konqueror.  The border and cell values must be specified as ="0" rather than ="0%" and a few other minrow tweaks to get this working across browsers.
Modified the header template so that it is used in the editcomment
screen (with the menu turned off).  Also modified the header so that
it looks more like Codestriker 1.7.8, with the black background and
the menu.

Fixed up the tab text so that they are shorter.

Modified the topic properties screen so that there are no alternating
colours, and so that it tabs correctly.