Merged changes from Jason:
Here is the long awaited patch that combines the topic history stuff
with the metric page. It was a bit more work than I expected. But it
turned out nice.

1. Codestriker calculates the "Codestriker Time" metric from the
topicviewtable. It is reported in both the property page, and the raw
metric download section.

2. The topichistory table is used to track the amount of time in each
state.

3. The topic history is displayed at the bottom of the info page.

4. Fixed the user metric report to report days from today, rather than
the absolute date.

5. Made the code that reports topic size ignore the context diff.

6. Made changes in the topic properties page to not change the changed
topic author to be == to the current users cookie.