Fixes from Jason:
The changes fix two other problems (both minor). First one, the
new comment metric tables was not being cleaned up when a topic was
deleted. The second problem is a missing escape on the codestriker
title. This will deal with a user overriding the title and include
html chars (minor, but I was writing a test set for this feature). The
third change is to allow the test cases to override the .conf file
without messing up the base config file. The new test scripts require
this change.  The two new test files are my first test scripts that
test codestriker with different configuration options. In the long
term really this should be the focus of the test scripts I wanted to
get the mechanism started. I am still not sure I am completly happy
with it, but it is simple and it works.