Comment tooltip now uses XMLHttpRequest object rather than using an iframe. Modified SubmitNewCommentAction so that it can send an XML reply, if that is what the format parameter requested.
The tooltip is dragged by Control-LeftMouseButton.  This needed to be changed
otherwise text could not be entered into the comment fields.  There may be
a way of have a "Drag" link, which can activate the drag, as on my Linux
box, the window manage grabs the Control-LeftMouseButton event, and drags
the entire browser window!